refactor(prompts): consolidate agent system instructions in one package
The chat agent's persona lived inline in the CLI chat handler (~100
lines) and the task-debugging persona inside the debug TUI, which made
them hard to find and iterate on together. Both now live in
internal/prompts, a dependency-free package importable from any layer.

The resource-recommendation prompts intentionally stay with their
generator: they are data-formatting templates, not standalone personas.
